Delete from S3 by using DeleteFromS3Worker to fix performance issue

+++ app/models/s3/util.rb
@@ -47,13 +47,13 @@ module S3
     def store(key, body, async: true)
       raise 'key is nil' if key.nil?
       ApplicationRecord.benchmark("#{self} Store by #{key} with async #{async}", level: :debug) do
-        cache.write(key.to_s, body)
-
         if async
           WriteToS3Worker.perform_async(klass: self, bucket: bucket_name, key: key.to_s, body: body)
         else
           client.put_object(bucket: bucket_name, key: key.to_s, body: body)
         end
+
+        cache.write(key.to_s, body)
       end
     end
 
@@ -66,9 +66,17 @@ module S3
       end
     end
 
-    def delete(key)
-      client.delete_object(bucket: bucket_name, key: key.to_s)
-      cache.delete(key.to_s)
+    def delete(key, async: true)
+      raise 'key is nil' if key.nil?
+      ApplicationRecord.benchmark("#{self} Delete by #{key} with async #{async}", level: :debug) do
+        if async
+          DeleteFromS3Worker.perform_async(klass: self, bucket: bucket_name, key: key.to_s)
+        else
+          client.delete_object(bucket: bucket_name, key: key.to_s)
+        end
+
+        cache.delete(key.to_s)
+      end
     end
 
     def exist(key)
